Fix findings across documentation and code
- AGENTS.md: Change 'parameterised' to 'parameterized' (en-GB-oxendict)
- Cargo.toml: Remove redundant serde_json from dev-dependencies
- normalize.rs: Extract where-clause parsing to new where_clause.rs module
- normalize.rs: Return not_implemented error for multi-focus arrays
- normalization_tests.rs: Clarify comment about where_clauses being discarded
- complexity-antipatterns.md: Add missing footnotes [^33] and [^34]
- phase-4-weaver-graph.md: Wrap trait methods in fenced rust code block
- jacquard-card-first-symbol-graph-design.md: Indent footnote continuations
- weaver-design.md: Fix invalid JSON example formatting
