+++ crates/sempai/Cargo.toml
@@ -14,7 +14,6 @@ sempai_core = { path = "../sempai-core", features = ["test-support"] }
 rstest = { workspace = true }
 rstest-bdd = { workspace = true }
 rstest-bdd-macros = { workspace = true }
-serde_json = { workspace = true }
 
 [lints]
 workspace = true
